requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pyngrok==7.0.5
python-dotenv==1.0.1
requests==2.31.0 